% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Account}
\alias{Account}
\title{Content API for Shopping Objects 
Manages product items, inventory, and Merchant Center accounts for Google Shopping.}
\usage{
Account(adultContent = NULL, adwordsLinks = NULL, id = NULL,
  name = NULL, reviewsUrl = NULL, sellerId = NULL, users = NULL,
  websiteUrl = NULL)
}
\arguments{
\item{adultContent}{Indicates whether the merchant sells adult content}

\item{adwordsLinks}{List of linked AdWords accounts that are active or pending approval}

\item{id}{Merchant Center account ID}

\item{name}{Display name for the account}

\item{reviewsUrl}{URL for individual seller reviews, i}

\item{sellerId}{Client-specific, locally-unique, internal ID for the child account}

\item{users}{Users with access to the account}

\item{websiteUrl}{The merchant's website}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:40:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlecontentv2.auto/R/content_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Account data.
}
\seealso{
Other Account functions: \code{\link{accounts.insert}},
  \code{\link{accounts.patch}},
  \code{\link{accounts.update}}
}
