% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountAdwordsLink}
\alias{AccountAdwordsLink}
\title{AccountAdwordsLink Object}
\usage{
AccountAdwordsLink(adwordsId = NULL, status = NULL)
}
\arguments{
\item{adwordsId}{Customer ID of the AdWords account}

\item{status}{Status of the link between this Merchant Center account and the AdWords account}
}
\value{
AccountAdwordsLink object
}
\description{
AccountAdwordsLink Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

