% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountIdentifier}
\alias{AccountIdentifier}
\title{AccountIdentifier Object}
\usage{
AccountIdentifier(aggregatorId = NULL, merchantId = NULL)
}
\arguments{
\item{aggregatorId}{The aggregator ID, set for aggregators and subaccounts (in that case, it represents the aggregator of the subaccount)}

\item{merchantId}{The merchant account ID, set for individual accounts and subaccounts}
}
\value{
AccountIdentifier object
}
\description{
AccountIdentifier Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

