% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{accounts.authinfo}
\alias{accounts.authinfo}
\title{Returns information about the authenticated user.}
\usage{
accounts.authinfo()
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details. 

#' @importFrom googleAuthR gar_api_generator
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}
