% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{accounts.custombatch}
\alias{accounts.custombatch}
\title{Retrieves, inserts, updates, and deletes multiple Merchant Center (sub-)accounts in a single request.}
\usage{
accounts.custombatch(AccountsCustomBatchRequest, dryRun = NULL)
}
\arguments{
\item{AccountsCustomBatchRequest}{The \link{AccountsCustomBatchRequest} object to pass to this method}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other AccountsCustomBatchRequest functions: \code{\link{AccountsCustomBatchRequest}}
}

