% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{accounts.delete}
\alias{accounts.delete}
\title{Deletes a Merchant Center sub-account.}
\usage{
accounts.delete(merchantId, accountId, dryRun = NULL)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{accountId}{The ID of the account}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}

