% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsAuthInfoResponse}
\alias{AccountsAuthInfoResponse}
\title{AccountsAuthInfoResponse Object}
\usage{
AccountsAuthInfoResponse(accountIdentifiers = NULL)
}
\arguments{
\item{accountIdentifiers}{The account identifiers corresponding to the authenticated user}
}
\value{
AccountsAuthInfoResponse object
}
\description{
AccountsAuthInfoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
