% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsCustomBatchRequest}
\alias{AccountsCustomBatchRequest}
\title{AccountsCustomBatchRequest Object}
\usage{
AccountsCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
AccountsCustomBatchRequest object
}
\description{
AccountsCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AccountsCustomBatchRequest functions: \code{\link{accounts.custombatch}}
}

