% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsCustomBatchRequestEntry}
\alias{AccountsCustomBatchRequestEntry}
\title{AccountsCustomBatchRequestEntry Object}
\usage{
AccountsCustomBatchRequestEntry(account = NULL, accountId = NULL,
  batchId = NULL, merchantId = NULL, method = NULL)
}
\arguments{
\item{account}{The account to create or update}

\item{accountId}{The ID of the account to get or delete}

\item{batchId}{An entry ID, unique within the batch request}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}
}
\value{
AccountsCustomBatchRequestEntry object
}
\description{
AccountsCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accounts request.
}
