% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsCustomBatchResponse}
\alias{AccountsCustomBatchResponse}
\title{AccountsCustomBatchResponse Object}
\usage{
AccountsCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
AccountsCustomBatchResponse object
}
\description{
AccountsCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
