% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsCustomBatchResponseEntry}
\alias{AccountsCustomBatchResponseEntry}
\title{AccountsCustomBatchResponseEntry Object}
\usage{
AccountsCustomBatchResponseEntry(account = NULL, batchId = NULL,
  errors = NULL)
}
\arguments{
\item{account}{The retrieved, created, or updated account}

\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}
}
\value{
AccountsCustomBatchResponseEntry object
}
\description{
AccountsCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accounts response.
}
