% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{accountshipping.get}
\alias{accountshipping.get}
\title{Retrieves the shipping settings of the account. This method can only be called for accounts to which the managing account has access: either the managing account itself or sub-accounts if the managing account is a multi-client account.}
\usage{
accountshipping.get(merchantId, accountId)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{accountId}{The ID of the account for which to get/update account shipping settings}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}
