% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShipping}
\alias{AccountShipping}
\title{AccountShipping Object}
\usage{
AccountShipping(accountId = NULL, carrierRates = NULL,
  locationGroups = NULL, rateTables = NULL, services = NULL)
}
\arguments{
\item{accountId}{The ID of the account to which these account shipping settings belong}

\item{carrierRates}{Carrier-based shipping calculations}

\item{locationGroups}{Location groups for shipping}

\item{rateTables}{Rate tables definitions}

\item{services}{Shipping services describing shipping fees calculation}
}
\value{
AccountShipping object
}
\description{
AccountShipping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The shipping settings of a merchant account.
}
\seealso{
Other AccountShipping functions: \code{\link{accountshipping.patch}},
  \code{\link{accountshipping.update}}
}

