% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingCondition}
\alias{AccountShippingCondition}
\title{AccountShippingCondition Object}
\usage{
AccountShippingCondition(deliveryLocationGroup = NULL,
  deliveryLocationId = NULL, deliveryPostalCode = NULL,
  deliveryPostalCodeRange = NULL, priceMax = NULL, shippingLabel = NULL,
  weightMax = NULL)
}
\arguments{
\item{deliveryLocationGroup}{Delivery location in terms of a location group name}

\item{deliveryLocationId}{Delivery location in terms of a location ID}

\item{deliveryPostalCode}{Delivery location in terms of a postal code}

\item{deliveryPostalCodeRange}{Delivery location in terms of a postal code range}

\item{priceMax}{Maximum shipping price}

\item{shippingLabel}{Shipping label of the product}

\item{weightMax}{Maximum shipping weight}
}
\value{
AccountShippingCondition object
}
\description{
AccountShippingCondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
