% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountshippingCustomBatchRequest}
\alias{AccountshippingCustomBatchRequest}
\title{AccountshippingCustomBatchRequest Object}
\usage{
AccountshippingCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
AccountshippingCustomBatchRequest object
}
\description{
AccountshippingCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AccountshippingCustomBatchRequest functions: \code{\link{accountshipping.custombatch}}
}
