% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountshippingCustomBatchRequestEntry}
\alias{AccountshippingCustomBatchRequestEntry}
\title{AccountshippingCustomBatchRequestEntry Object}
\usage{
AccountshippingCustomBatchRequestEntry(accountId = NULL,
  accountShipping = NULL, batchId = NULL, merchantId = NULL,
  method = NULL)
}
\arguments{
\item{accountId}{The ID of the account for which to get/update account shipping settings}

\item{accountShipping}{The account shipping settings to update}

\item{batchId}{An entry ID, unique within the batch request}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}
}
\value{
AccountshippingCustomBatchRequestEntry object
}
\description{
AccountshippingCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accountshipping request.
}
