% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountshippingCustomBatchResponse}
\alias{AccountshippingCustomBatchResponse}
\title{AccountshippingCustomBatchResponse Object}
\usage{
AccountshippingCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
AccountshippingCustomBatchResponse object
}
\description{
AccountshippingCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

