% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountshippingCustomBatchResponseEntry}
\alias{AccountshippingCustomBatchResponseEntry}
\title{AccountshippingCustomBatchResponseEntry Object}
\usage{
AccountshippingCustomBatchResponseEntry(accountShipping = NULL,
  batchId = NULL, errors = NULL)
}
\arguments{
\item{accountShipping}{The retrieved or updated account shipping settings}

\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}
}
\value{
AccountshippingCustomBatchResponseEntry object
}
\description{
AccountshippingCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accountshipping response.
}

