% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountshippingListResponse}
\alias{AccountshippingListResponse}
\title{AccountshippingListResponse Object}
\usage{
AccountshippingListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of account shipping settings}

\item{resources}{No description}
}
\value{
AccountshippingListResponse object
}
\description{
AccountshippingListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
