% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingLocationGroup}
\alias{AccountShippingLocationGroup}
\title{AccountShippingLocationGroup Object}
\usage{
AccountShippingLocationGroup(country = NULL, locationIds = NULL,
  name = NULL, postalCodeRanges = NULL, postalCodes = NULL)
}
\arguments{
\item{country}{The CLDR territory code of the country in which this location group is}

\item{locationIds}{A location ID (also called criteria ID) representing administrative areas, smaller country subdivisions (counties), or cities}

\item{name}{The name of the location group}

\item{postalCodeRanges}{A postal code range representing a city or a set of cities}

\item{postalCodes}{A postal code representing a city or a set of cities}
}
\value{
AccountShippingLocationGroup object
}
\description{
AccountShippingLocationGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A user-defined locations group in a given country. All the locations of the group must be of the same type.
}
