% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingPostalCodeRange}
\alias{AccountShippingPostalCodeRange}
\title{AccountShippingPostalCodeRange Object}
\usage{
AccountShippingPostalCodeRange(end = NULL, start = NULL)
}
\arguments{
\item{end}{The last (inclusive) postal code or prefix of the range}

\item{start}{The first (inclusive) postal code or prefix of the range}
}
\value{
AccountShippingPostalCodeRange object
}
\description{
AccountShippingPostalCodeRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A postal code range, that can be either:  - A range of postal codes (e.g., start=12340, end=12359)- A range of postal codes prefixes (e.g., start=1234* end=1235*). Prefixes must be of the same length (e.g., start=12* end=2* is invalid).
}

