% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingRateTable}
\alias{AccountShippingRateTable}
\title{AccountShippingRateTable Object}
\usage{
AccountShippingRateTable(content = NULL, name = NULL, saleCountry = NULL)
}
\arguments{
\item{content}{One-dimensional table cells define one condition along the same dimension}

\item{name}{The name of the rate table}

\item{saleCountry}{The sale country for which this table is valid, represented as a CLDR territory code}
}
\value{
AccountShippingRateTable object
}
\description{
AccountShippingRateTable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single or bi-dimensional table of shipping rates. Each dimension is defined in terms of consecutive price/weight ranges, delivery locations, or shipping labels.
}
