% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingRateTableCell}
\alias{AccountShippingRateTableCell}
\title{AccountShippingRateTableCell Object}
\usage{
AccountShippingRateTableCell(condition = NULL, rate = NULL)
}
\arguments{
\item{condition}{Conditions for which the cell is valid}

\item{rate}{The rate applicable if the cell conditions are matched}
}
\value{
AccountShippingRateTableCell object
}
\description{
AccountShippingRateTableCell Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
