% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingShippingService}
\alias{AccountShippingShippingService}
\title{AccountShippingShippingService Object}
\usage{
AccountShippingShippingService(active = NULL, calculationMethod = NULL,
  costRuleTree = NULL, maxDaysInTransit = NULL, minDaysInTransit = NULL,
  name = NULL, saleCountry = NULL)
}
\arguments{
\item{active}{Whether the shipping service is available}

\item{calculationMethod}{Calculation method for the 'simple' case that needs no rules}

\item{costRuleTree}{Decision tree for 'complicated' shipping cost calculation}

\item{maxDaysInTransit}{The maximum number of days in transit}

\item{minDaysInTransit}{The minimum number of days in transit}

\item{name}{The name of this shipping service}

\item{saleCountry}{The CLDR territory code of the sale country for which this service can be used}
}
\value{
AccountShippingShippingService object
}
\description{
AccountShippingShippingService Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Shipping services provided in a country.
}
