% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountShippingShippingServiceCostRule}
\alias{AccountShippingShippingServiceCostRule}
\title{AccountShippingShippingServiceCostRule Object}
\usage{
AccountShippingShippingServiceCostRule(calculationMethod = NULL,
  children = NULL, condition = NULL)
}
\arguments{
\item{calculationMethod}{Final calculation method to be used only in leaf nodes}

\item{children}{Subsequent rules to be applied, only for inner nodes}

\item{condition}{Condition for this rule to be applicable}
}
\value{
AccountShippingShippingServiceCostRule object
}
\description{
AccountShippingShippingServiceCostRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Building block of the cost calculation decision tree.  - The tree root should have no condition and no calculation method.- All the children must have a condition on the same dimension. The first child matching a condition is entered, therefore, price and weight conditions form contiguous intervals.- The last child of an element must have no condition and matches all elements not previously matched.- Children and calculation method are mutually exclusive, and exactly one of them must be defined; the root must only have children.
}

