% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountsListResponse}
\alias{AccountsListResponse}
\title{AccountsListResponse Object}
\usage{
AccountsListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of accounts}

\item{resources}{No description}
}
\value{
AccountsListResponse object
}
\description{
AccountsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

