% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountStatus}
\alias{AccountStatus}
\title{AccountStatus Object}
\usage{
AccountStatus(accountId = NULL, dataQualityIssues = NULL)
}
\arguments{
\item{accountId}{The ID of the account for which the status is reported}

\item{dataQualityIssues}{A list of data quality issues}
}
\value{
AccountStatus object
}
\description{
AccountStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The status of an account, i.e., information about its products, which is computed offline and not returned immediately at insertion time.
}

