% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountStatusDataQualityIssue}
\alias{AccountStatusDataQualityIssue}
\title{AccountStatusDataQualityIssue Object}
\usage{
AccountStatusDataQualityIssue(country = NULL, detail = NULL,
  displayedValue = NULL, exampleItems = NULL, id = NULL,
  lastChecked = NULL, location = NULL, numItems = NULL, severity = NULL,
  submittedValue = NULL)
}
\arguments{
\item{country}{Country for which this issue is reported}

\item{detail}{A more detailed description of the issue}

\item{displayedValue}{Actual value displayed on the landing page}

\item{exampleItems}{Example items featuring the issue}

\item{id}{Issue identifier}

\item{lastChecked}{Last time the account was checked for this issue}

\item{location}{The attribute name that is relevant for the issue}

\item{numItems}{Number of items in the account found to have the said issue}

\item{severity}{Severity of the problem}

\item{submittedValue}{Submitted value that causes the issue}
}
\value{
AccountStatusDataQualityIssue object
}
\description{
AccountStatusDataQualityIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
