% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountstatusesCustomBatchRequest}
\alias{AccountstatusesCustomBatchRequest}
\title{AccountstatusesCustomBatchRequest Object}
\usage{
AccountstatusesCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
AccountstatusesCustomBatchRequest object
}
\description{
AccountstatusesCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AccountstatusesCustomBatchRequest functions: \code{\link{accountstatuses.custombatch}}
}

