% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountstatusesCustomBatchResponse}
\alias{AccountstatusesCustomBatchResponse}
\title{AccountstatusesCustomBatchResponse Object}
\usage{
AccountstatusesCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
AccountstatusesCustomBatchResponse object
}
\description{
AccountstatusesCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
