% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountstatusesCustomBatchResponseEntry}
\alias{AccountstatusesCustomBatchResponseEntry}
\title{AccountstatusesCustomBatchResponseEntry Object}
\usage{
AccountstatusesCustomBatchResponseEntry(accountStatus = NULL,
  batchId = NULL, errors = NULL)
}
\arguments{
\item{accountStatus}{The requested account status}

\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}
}
\value{
AccountstatusesCustomBatchResponseEntry object
}
\description{
AccountstatusesCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accountstatuses response.
}

