% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountstatusesListResponse}
\alias{AccountstatusesListResponse}
\title{AccountstatusesListResponse Object}
\usage{
AccountstatusesListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of account statuses}

\item{resources}{No description}
}
\value{
AccountstatusesListResponse object
}
\description{
AccountstatusesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
