% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountStatusExampleItem}
\alias{AccountStatusExampleItem}
\title{AccountStatusExampleItem Object}
\usage{
AccountStatusExampleItem(itemId = NULL, link = NULL,
  submittedValue = NULL, title = NULL, valueOnLandingPage = NULL)
}
\arguments{
\item{itemId}{Unique item ID as specified in the uploaded product data}

\item{link}{Landing page of the item}

\item{submittedValue}{The item value that was submitted}

\item{title}{Title of the item}

\item{valueOnLandingPage}{The actual value on the landing page}
}
\value{
AccountStatusExampleItem object
}
\description{
AccountStatusExampleItem Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An example of an item that has poor data quality. An item value on the landing page differs from what is submitted, or conflicts with a policy.
}

