% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountTax}
\alias{AccountTax}
\title{AccountTax Object}
\usage{
AccountTax(accountId = NULL, rules = NULL)
}
\arguments{
\item{accountId}{The ID of the account to which these account tax settings belong}

\item{rules}{Tax rules}
}
\value{
AccountTax object
}
\description{
AccountTax Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The tax settings of a merchant account.
}
\seealso{
Other AccountTax functions: \code{\link{accounttax.patch}},
  \code{\link{accounttax.update}}
}

