% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccounttaxCustomBatchRequest}
\alias{AccounttaxCustomBatchRequest}
\title{AccounttaxCustomBatchRequest Object}
\usage{
AccounttaxCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
AccounttaxCustomBatchRequest object
}
\description{
AccounttaxCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other AccounttaxCustomBatchRequest functions: \code{\link{accounttax.custombatch}}
}
