% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccounttaxCustomBatchRequestEntry}
\alias{AccounttaxCustomBatchRequestEntry}
\title{AccounttaxCustomBatchRequestEntry Object}
\usage{
AccounttaxCustomBatchRequestEntry(accountId = NULL, accountTax = NULL,
  batchId = NULL, merchantId = NULL, method = NULL)
}
\arguments{
\item{accountId}{The ID of the account for which to get/update account tax settings}

\item{accountTax}{The account tax settings to update}

\item{batchId}{An entry ID, unique within the batch request}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}
}
\value{
AccounttaxCustomBatchRequestEntry object
}
\description{
AccounttaxCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accounttax request.
}

