% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccounttaxCustomBatchResponse}
\alias{AccounttaxCustomBatchResponse}
\title{AccounttaxCustomBatchResponse Object}
\usage{
AccounttaxCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
AccounttaxCustomBatchResponse object
}
\description{
AccounttaxCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

