% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccounttaxCustomBatchResponseEntry}
\alias{AccounttaxCustomBatchResponseEntry}
\title{AccounttaxCustomBatchResponseEntry Object}
\usage{
AccounttaxCustomBatchResponseEntry(accountTax = NULL, batchId = NULL,
  errors = NULL)
}
\arguments{
\item{accountTax}{The retrieved or updated account tax settings}

\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}
}
\value{
AccounttaxCustomBatchResponseEntry object
}
\description{
AccounttaxCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accounttax response.
}

