% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccounttaxListResponse}
\alias{AccounttaxListResponse}
\title{AccounttaxListResponse Object}
\usage{
AccounttaxListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of account tax settings}

\item{resources}{No description}
}
\value{
AccounttaxListResponse object
}
\description{
AccounttaxListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

