% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountTaxTaxRule}
\alias{AccountTaxTaxRule}
\title{AccountTaxTaxRule Object}
\usage{
AccountTaxTaxRule(country = NULL, locationId = NULL, ratePercent = NULL,
  shippingTaxed = NULL, useGlobalRate = NULL)
}
\arguments{
\item{country}{Country code in which tax is applicable}

\item{locationId}{State (or province) is which the tax is applicable, described by its location id (also called criteria id)}

\item{ratePercent}{Explicit tax rate in percent, represented as a floating point number without the percentage character}

\item{shippingTaxed}{If true, shipping charges are also taxed}

\item{useGlobalRate}{Whether the tax rate is taken from a global tax table or specified explicitly}
}
\value{
AccountTaxTaxRule object
}
\description{
AccountTaxTaxRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Tax calculation rule to apply in a state or province (USA only).
}

