% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{AccountUser}
\alias{AccountUser}
\title{AccountUser Object}
\usage{
AccountUser(admin = NULL, emailAddress = NULL)
}
\arguments{
\item{admin}{Whether user is an admin}

\item{emailAddress}{User's email address}
}
\value{
AccountUser object
}
\description{
AccountUser Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
