% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{CarrierRate}
\alias{CarrierRate}
\title{CarrierRate Object}
\usage{
CarrierRate(carrierName = NULL, carrierService = NULL,
  flatAdjustment = NULL, name = NULL, originPostalCode = NULL,
  percentageAdjustment = NULL)
}
\arguments{
\item{carrierName}{Carrier service, such as 'UPS' or 'Fedex'}

\item{carrierService}{Carrier service, such as 'ground' or '2 days'}

\item{flatAdjustment}{Additive shipping rate modifier}

\item{name}{Name of the carrier rate}

\item{originPostalCode}{Shipping origin for this carrier rate}

\item{percentageAdjustment}{Multiplicative shipping rate modifier as a number in decimal notation}
}
\value{
CarrierRate object
}
\description{
CarrierRate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

