% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{CarriersCarrier}
\alias{CarriersCarrier}
\title{CarriersCarrier Object}
\usage{
CarriersCarrier(country = NULL, name = NULL, services = NULL)
}
\arguments{
\item{country}{The CLDR country code of the carrier (e}

\item{name}{The name of the carrier (e}

\item{services}{A list of supported services (e}
}
\value{
CarriersCarrier object
}
\description{
CarriersCarrier Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

