% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Datafeed}
\alias{Datafeed}
\title{Datafeed Object}
\usage{
Datafeed(attributeLanguage = NULL, contentLanguage = NULL,
  contentType = NULL, fetchSchedule = NULL, fileName = NULL,
  format = NULL, id = NULL, intendedDestinations = NULL, name = NULL,
  targetCountry = NULL)
}
\arguments{
\item{attributeLanguage}{The two-letter ISO 639-1 language in which the attributes are defined in the data feed}

\item{contentLanguage}{The two-letter ISO 639-1 language of the items in the feed}

\item{contentType}{The type of data feed}

\item{fetchSchedule}{Fetch schedule for the feed file}

\item{fileName}{The filename of the feed}

\item{format}{Format of the feed file}

\item{id}{The ID of the data feed}

\item{intendedDestinations}{The list of intended destinations (corresponds to checked check boxes in Merchant Center)}

\item{name}{A descriptive name of the data feed}

\item{targetCountry}{The country where the items in the feed will be included in the search index, represented as a CLDR territory code}
}
\value{
Datafeed object
}
\description{
Datafeed Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Datafeed data.
}
\seealso{
Other Datafeed functions: \code{\link{datafeeds.insert}},
  \code{\link{datafeeds.patch}},
  \code{\link{datafeeds.update}}
}

