% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedFetchSchedule}
\alias{DatafeedFetchSchedule}
\title{DatafeedFetchSchedule Object}
\usage{
DatafeedFetchSchedule(dayOfMonth = NULL, fetchUrl = NULL, hour = NULL,
  minuteOfHour = NULL, password = NULL, timeZone = NULL,
  username = NULL, weekday = NULL)
}
\arguments{
\item{dayOfMonth}{The day of the month the feed file should be fetched (1-31)}

\item{fetchUrl}{The URL where the feed file can be fetched}

\item{hour}{The hour of the day the feed file should be fetched (0-23)}

\item{minuteOfHour}{The minute of the hour the feed file should be fetched (0-59)}

\item{password}{An optional password for fetch_url}

\item{timeZone}{Time zone used for schedule}

\item{username}{An optional user name for fetch_url}

\item{weekday}{The day of the week the feed file should be fetched}
}
\value{
DatafeedFetchSchedule object
}
\description{
DatafeedFetchSchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The required fields vary based on the frequency of fetching. For a monthly fetch schedule, day_of_month and hour are required. For a weekly fetch schedule, weekday and hour are required. For a daily fetch schedule, only hour is required.
}
