% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedFormat}
\alias{DatafeedFormat}
\title{DatafeedFormat Object}
\usage{
DatafeedFormat(columnDelimiter = NULL, fileEncoding = NULL,
  quotingMode = NULL)
}
\arguments{
\item{columnDelimiter}{Delimiter for the separation of values in a delimiter-separated values feed}

\item{fileEncoding}{Character encoding scheme of the data feed}

\item{quotingMode}{Specifies how double quotes are interpreted}
}
\value{
DatafeedFormat object
}
\description{
DatafeedFormat Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
