% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{datafeeds.custombatch}
\alias{datafeeds.custombatch}
\title{Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}}
\usage{
datafeeds.custombatch(DatafeedsCustomBatchRequest, dryRun = NULL)
}
\arguments{
\item{DatafeedsCustomBatchRequest}{The \link{DatafeedsCustomBatchRequest} object to pass to this method}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other DatafeedsCustomBatchRequest functions: \code{\link{DatafeedsCustomBatchRequest}}
}
