% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{datafeeds.insert}
\alias{datafeeds.insert}
\title{Registers a datafeed with your Merchant Center account. This method can only be called for non-multi-client accounts.}
\usage{
datafeeds.insert(Datafeed, merchantId, dryRun = NULL)
}
\arguments{
\item{Datafeed}{The \link{Datafeed} object to pass to this method}

\item{merchantId}{}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other Datafeed functions: \code{\link{Datafeed}},
  \code{\link{datafeeds.patch}},
  \code{\link{datafeeds.update}}
}
