% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedsCustomBatchRequest}
\alias{DatafeedsCustomBatchRequest}
\title{DatafeedsCustomBatchRequest Object}
\usage{
DatafeedsCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
DatafeedsCustomBatchRequest object
}
\description{
DatafeedsCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DatafeedsCustomBatchRequest functions: \code{\link{datafeeds.custombatch}}
}
