% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedsCustomBatchRequestEntry}
\alias{DatafeedsCustomBatchRequestEntry}
\title{DatafeedsCustomBatchRequestEntry Object}
\usage{
DatafeedsCustomBatchRequestEntry(batchId = NULL, datafeed = NULL,
  datafeedId = NULL, merchantId = NULL, method = NULL)
}
\arguments{
\item{batchId}{An entry ID, unique within the batch request}

\item{datafeed}{The data feed to insert}

\item{datafeedId}{The ID of the data feed to get or delete}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}
}
\value{
DatafeedsCustomBatchRequestEntry object
}
\description{
DatafeedsCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch datafeeds request.
}
