% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedsCustomBatchResponse}
\alias{DatafeedsCustomBatchResponse}
\title{DatafeedsCustomBatchResponse Object}
\usage{
DatafeedsCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
DatafeedsCustomBatchResponse object
}
\description{
DatafeedsCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
