% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedsListResponse}
\alias{DatafeedsListResponse}
\title{DatafeedsListResponse Object}
\usage{
DatafeedsListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of datafeeds}

\item{resources}{No description}
}
\value{
DatafeedsListResponse object
}
\description{
DatafeedsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
