% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedStatus}
\alias{DatafeedStatus}
\title{DatafeedStatus Object}
\usage{
DatafeedStatus(datafeedId = NULL, errors = NULL, itemsTotal = NULL,
  itemsValid = NULL, lastUploadDate = NULL, processingStatus = NULL,
  warnings = NULL)
}
\arguments{
\item{datafeedId}{The ID of the feed for which the status is reported}

\item{errors}{The list of errors occurring in the feed}

\item{itemsTotal}{The number of items in the feed that were processed}

\item{itemsValid}{The number of items in the feed that were valid}

\item{lastUploadDate}{The last date at which the feed was uploaded}

\item{processingStatus}{The processing status of the feed}

\item{warnings}{The list of errors occurring in the feed}
}
\value{
DatafeedStatus object
}
\description{
DatafeedStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The status of a datafeed, i.e., the result of the last retrieval of the datafeed computed asynchronously when the feed processing is finished.
}
