% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedStatusError}
\alias{DatafeedStatusError}
\title{DatafeedStatusError Object}
\usage{
DatafeedStatusError(code = NULL, count = NULL, examples = NULL,
  message = NULL)
}
\arguments{
\item{code}{The code of the error, e}

\item{count}{The number of occurrences of the error in the feed}

\item{examples}{A list of example occurrences of the error, grouped by product}

\item{message}{The error message, e}
}
\value{
DatafeedStatusError object
}
\description{
DatafeedStatusError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error occurring in the feed, like 'invalid price'.
}

