% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedstatusesCustomBatchRequest}
\alias{DatafeedstatusesCustomBatchRequest}
\title{DatafeedstatusesCustomBatchRequest Object}
\usage{
DatafeedstatusesCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
DatafeedstatusesCustomBatchRequest object
}
\description{
DatafeedstatusesCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other DatafeedstatusesCustomBatchRequest functions: \code{\link{datafeedstatuses.custombatch}}
}

