% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedstatusesCustomBatchRequestEntry}
\alias{DatafeedstatusesCustomBatchRequestEntry}
\title{DatafeedstatusesCustomBatchRequestEntry Object}
\usage{
DatafeedstatusesCustomBatchRequestEntry(batchId = NULL, datafeedId = NULL,
  merchantId = NULL, method = NULL)
}
\arguments{
\item{batchId}{An entry ID, unique within the batch request}

\item{datafeedId}{The ID of the data feed to get or delete}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}
}
\value{
DatafeedstatusesCustomBatchRequestEntry object
}
\description{
DatafeedstatusesCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch datafeedstatuses request.
}

