% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedstatusesCustomBatchResponse}
\alias{DatafeedstatusesCustomBatchResponse}
\title{DatafeedstatusesCustomBatchResponse Object}
\usage{
DatafeedstatusesCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
DatafeedstatusesCustomBatchResponse object
}
\description{
DatafeedstatusesCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
