% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedstatusesCustomBatchResponseEntry}
\alias{DatafeedstatusesCustomBatchResponseEntry}
\title{DatafeedstatusesCustomBatchResponseEntry Object}
\usage{
DatafeedstatusesCustomBatchResponseEntry(batchId = NULL,
  datafeedStatus = NULL, errors = NULL)
}
\arguments{
\item{batchId}{The ID of the request entry this entry responds to}

\item{datafeedStatus}{The requested data feed status}

\item{errors}{A list of errors defined if and only if the request failed}
}
\value{
DatafeedstatusesCustomBatchResponseEntry object
}
\description{
DatafeedstatusesCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch datafeedstatuses response.
}
