% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedstatusesListResponse}
\alias{DatafeedstatusesListResponse}
\title{DatafeedstatusesListResponse Object}
\usage{
DatafeedstatusesListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of datafeed statuses}

\item{resources}{No description}
}
\value{
DatafeedstatusesListResponse object
}
\description{
DatafeedstatusesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

