% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DatafeedStatusExample}
\alias{DatafeedStatusExample}
\title{DatafeedStatusExample Object}
\usage{
DatafeedStatusExample(itemId = NULL, lineNumber = NULL, value = NULL)
}
\arguments{
\item{itemId}{The ID of the example item}

\item{lineNumber}{Line number in the data feed where the example is found}

\item{value}{The problematic value}
}
\value{
DatafeedStatusExample object
}
\description{
DatafeedStatusExample Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An example occurrence for a particular error.
}

