% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{DeliveryTime}
\alias{DeliveryTime}
\title{DeliveryTime Object}
\usage{
DeliveryTime(maxTransitTimeInDays = NULL, minTransitTimeInDays = NULL)
}
\arguments{
\item{maxTransitTimeInDays}{Maximum number of business days that is spent in transit}

\item{minTransitTimeInDays}{Minimum number of business days that is spent in transit}
}
\value{
DeliveryTime object
}
\description{
DeliveryTime Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
