% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Error}
\alias{Error}
\title{Error Object}
\usage{
Error(domain = NULL, message = NULL, reason = NULL)
}
\arguments{
\item{domain}{The domain of the error}

\item{message}{A description of the error}

\item{reason}{The error code}
}
\value{
Error object
}
\description{
Error Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error returned by the API.
}
