% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Headers}
\alias{Headers}
\title{Headers Object}
\usage{
Headers(locations = NULL, numberOfItems = NULL,
  postalCodeGroupNames = NULL, prices = NULL, weights = NULL)
}
\arguments{
\item{locations}{A list of location ID sets}

\item{numberOfItems}{A list of inclusive number of items upper bounds}

\item{postalCodeGroupNames}{A list of postal group names}

\item{prices}{be 'infinity'}

\item{weights}{be 'infinity'}
}
\value{
Headers object
}
\description{
Headers Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A non-empty list of row or column headers for a table. Exactly one of prices, weights, numItems, postalCodeGroupNames, or locations must be set.
}
