% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Installment}
\alias{Installment}
\title{Installment Object}
\usage{
Installment(amount = NULL, months = NULL)
}
\arguments{
\item{amount}{The amount the buyer has to pay per month}

\item{months}{The number of installments the buyer has to pay}
}
\value{
Installment object
}
\description{
Installment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

