% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{inventory.custombatch}
\alias{inventory.custombatch}
\title{Updates price and availability for multiple products or stores in a single request. This operation does not update the expiration date of the products. This method can only be called for non-multi-client accounts.}
\usage{
inventory.custombatch(InventoryCustomBatchRequest, dryRun = NULL)
}
\arguments{
\item{InventoryCustomBatchRequest}{The \link{InventoryCustomBatchRequest} object to pass to this method}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other InventoryCustomBatchRequest functions: \code{\link{InventoryCustomBatchRequest}}
}
