% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{inventory.set}
\alias{inventory.set}
\title{Updates price and availability of a product in your Merchant Center account. This operation does not update the expiration date of the product. This method can only be called for non-multi-client accounts.}
\usage{
inventory.set(InventorySetRequest, merchantId, storeCode, productId,
  dryRun = NULL)
}
\arguments{
\item{InventorySetRequest}{The \link{InventorySetRequest} object to pass to this method}

\item{merchantId}{The ID of the managing account}

\item{storeCode}{The code of the store for which to update price and availability}

\item{productId}{The ID of the product for which to update price and availability}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other InventorySetRequest functions: \code{\link{InventorySetRequest}}
}
