% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{InventoryCustomBatchRequest}
\alias{InventoryCustomBatchRequest}
\title{InventoryCustomBatchRequest Object}
\usage{
InventoryCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
InventoryCustomBatchRequest object
}
\description{
InventoryCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other InventoryCustomBatchRequest functions: \code{\link{inventory.custombatch}}
}
