% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{InventoryCustomBatchRequestEntry}
\alias{InventoryCustomBatchRequestEntry}
\title{InventoryCustomBatchRequestEntry Object}
\usage{
InventoryCustomBatchRequestEntry(batchId = NULL, inventory = NULL,
  merchantId = NULL, productId = NULL, storeCode = NULL)
}
\arguments{
\item{batchId}{An entry ID, unique within the batch request}

\item{inventory}{Price and availability of the product}

\item{merchantId}{The ID of the managing account}

\item{productId}{The ID of the product for which to update price and availability}

\item{storeCode}{The code of the store for which to update price and availability}
}
\value{
InventoryCustomBatchRequestEntry object
}
\description{
InventoryCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch inventory request.
}

