% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{InventoryCustomBatchResponse}
\alias{InventoryCustomBatchResponse}
\title{InventoryCustomBatchResponse Object}
\usage{
InventoryCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
InventoryCustomBatchResponse object
}
\description{
InventoryCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
