% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{InventoryPickup}
\alias{InventoryPickup}
\title{InventoryPickup Object}
\usage{
InventoryPickup(pickupMethod = NULL, pickupSla = NULL)
}
\arguments{
\item{pickupMethod}{Whether store pickup is available for this offer and whether the pickup option should be shown as buy, reserve, or not supported}

\item{pickupSla}{The expected date that an order will be ready for pickup, relative to when the order is placed}
}
\value{
InventoryPickup object
}
\description{
InventoryPickup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
