% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{LocationIdSet}
\alias{LocationIdSet}
\title{LocationIdSet Object}
\usage{
LocationIdSet(locationIds = NULL)
}
\arguments{
\item{locationIds}{A non-empty list of location IDs}
}
\value{
LocationIdSet object
}
\description{
LocationIdSet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

