% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{LoyaltyPoints}
\alias{LoyaltyPoints}
\title{LoyaltyPoints Object}
\usage{
LoyaltyPoints(name = NULL, pointsValue = NULL, ratio = NULL)
}
\arguments{
\item{name}{Name of loyalty points program}

\item{pointsValue}{The retailer's loyalty points in absolute value}

\item{ratio}{The ratio of a point when converted to currency}
}
\value{
LoyaltyPoints object
}
\description{
LoyaltyPoints Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

