% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Order}
\alias{Order}
\title{Order Object}
\usage{
Order(acknowledged = NULL, channelType = NULL, customer = NULL,
  deliveryDetails = NULL, id = NULL, lineItems = NULL,
  merchantId = NULL, merchantOrderId = NULL, netAmount = NULL,
  paymentMethod = NULL, paymentStatus = NULL, placedDate = NULL,
  promotions = NULL, refunds = NULL, shipments = NULL,
  shippingCost = NULL, shippingCostTax = NULL, shippingOption = NULL,
  status = NULL)
}
\arguments{
\item{acknowledged}{Whether the order was acknowledged}

\item{channelType}{The channel type of the order: 'purchaseOnGoogle' or 'googleExpress'}

\item{customer}{The details of the customer who placed the order}

\item{deliveryDetails}{The details for the delivery}

\item{id}{The REST id of the order}

\item{lineItems}{Line items that are ordered}

\item{merchantId}{No description}

\item{merchantOrderId}{Merchant-provided id of the order}

\item{netAmount}{The net amount for the order}

\item{paymentMethod}{The details of the payment method}

\item{paymentStatus}{The status of the payment}

\item{placedDate}{The date when the order was placed, in ISO 8601 format}

\item{promotions}{The details of the merchant provided promotions applied to the order}

\item{refunds}{Refunds for the order}

\item{shipments}{Shipments of the order}

\item{shippingCost}{The total cost of shipping for all items}

\item{shippingCostTax}{The tax for the total shipping cost}

\item{shippingOption}{The requested shipping option}

\item{status}{The status of the order}
}
\value{
Order object
}
\description{
Order Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
