% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderAddress}
\alias{OrderAddress}
\title{OrderAddress Object}
\usage{
OrderAddress(country = NULL, fullAddress = NULL, isPostOfficeBox = NULL,
  locality = NULL, postalCode = NULL, recipientName = NULL,
  region = NULL, streetAddress = NULL)
}
\arguments{
\item{country}{CLDR country code (e}

\item{fullAddress}{Strings representing the lines of the printed label for mailing the order, for example:}

\item{isPostOfficeBox}{Whether the address is a post office box}

\item{locality}{City, town or commune}

\item{postalCode}{Postal Code or ZIP (e}

\item{recipientName}{Name of the recipient}

\item{region}{Top-level administrative subdivision of the country (e}

\item{streetAddress}{Street-level part of the address}
}
\value{
OrderAddress object
}
\description{
OrderAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
