% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderCancellation}
\alias{OrderCancellation}
\title{OrderCancellation Object}
\usage{
OrderCancellation(actor = NULL, creationDate = NULL, quantity = NULL,
  reason = NULL, reasonText = NULL)
}
\arguments{
\item{actor}{The actor that created the cancellation}

\item{creationDate}{Date on which the cancellation has been created, in ISO 8601 format}

\item{quantity}{The quantity that was canceled}

\item{reason}{The reason for the cancellation}

\item{reasonText}{The explanation of the reason}
}
\value{
OrderCancellation object
}
\description{
OrderCancellation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
