% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderCustomer}
\alias{OrderCustomer}
\title{OrderCustomer Object}
\usage{
OrderCustomer(email = NULL, explicitMarketingPreference = NULL,
  fullName = NULL)
}
\arguments{
\item{email}{Email address of the customer}

\item{explicitMarketingPreference}{If set, this indicates the user explicitly chose to opt in or out of providing marketing rights to the merchant}

\item{fullName}{Full name of the customer}
}
\value{
OrderCustomer object
}
\description{
OrderCustomer Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

