% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderDeliveryDetails}
\alias{OrderDeliveryDetails}
\title{OrderDeliveryDetails Object}
\usage{
OrderDeliveryDetails(address = NULL, phoneNumber = NULL)
}
\arguments{
\item{address}{The delivery address}

\item{phoneNumber}{The phone number of the person receiving the delivery}
}
\value{
OrderDeliveryDetails object
}
\description{
OrderDeliveryDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

