% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderLineItemReturnInfo}
\alias{OrderLineItemReturnInfo}
\title{OrderLineItemReturnInfo Object}
\usage{
OrderLineItemReturnInfo(daysToReturn = NULL, isReturnable = NULL,
  policyUrl = NULL)
}
\arguments{
\item{daysToReturn}{How many days later the item can be returned}

\item{isReturnable}{Whether the item is returnable}

\item{policyUrl}{URL of the item return policy}
}
\value{
OrderLineItemReturnInfo object
}
\description{
OrderLineItemReturnInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

