% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderLineItemShippingDetails}
\alias{OrderLineItemShippingDetails}
\title{OrderLineItemShippingDetails Object}
\usage{
OrderLineItemShippingDetails(deliverByDate = NULL, method = NULL,
  shipByDate = NULL)
}
\arguments{
\item{deliverByDate}{The delivery by date, in ISO 8601 format}

\item{method}{Details of the shipping method}

\item{shipByDate}{The ship by date, in ISO 8601 format}
}
\value{
OrderLineItemShippingDetails object
}
\description{
OrderLineItemShippingDetails Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
