% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderRefund}
\alias{OrderRefund}
\title{OrderRefund Object}
\usage{
OrderRefund(actor = NULL, amount = NULL, creationDate = NULL,
  reason = NULL, reasonText = NULL)
}
\arguments{
\item{actor}{The actor that created the refund}

\item{amount}{The amount that is refunded}

\item{creationDate}{Date on which the item has been created, in ISO 8601 format}

\item{reason}{The reason for the refund}

\item{reasonText}{The explanation of the reason}
}
\value{
OrderRefund object
}
\description{
OrderRefund Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

