% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{orders.acknowledge}
\alias{orders.acknowledge}
\title{Marks an order as acknowledged. This method can only be called for non-multi-client accounts.}
\usage{
orders.acknowledge(OrdersAcknowledgeRequest, merchantId, orderId)
}
\arguments{
\item{OrdersAcknowledgeRequest}{The \link{OrdersAcknowledgeRequest} object to pass to this method}

\item{merchantId}{The ID of the managing account}

\item{orderId}{The ID of the order}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other OrdersAcknowledgeRequest functions: \code{\link{OrdersAcknowledgeRequest}}
}
