% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{orders.advancetestorder}
\alias{orders.advancetestorder}
\title{Sandbox only. Moves a test order from state 'inProgress' to state 'pendingShipment'.}
\usage{
orders.advancetestorder(merchantId, orderId)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{orderId}{The ID of the test order to modify}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}

