% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{orders.gettestordertemplate}
\alias{orders.gettestordertemplate}
\title{Sandbox only. Retrieves an order template that can be used to quickly create a new order in sandbox.}
\usage{
orders.gettestordertemplate(merchantId, templateName)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{templateName}{The name of the template to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}

