% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersAcknowledgeResponse}
\alias{OrdersAcknowledgeResponse}
\title{OrdersAcknowledgeResponse Object}
\usage{
OrdersAcknowledgeResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersAcknowledgeResponse object
}
\description{
OrdersAcknowledgeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

