% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCancelLineItemResponse}
\alias{OrdersCancelLineItemResponse}
\title{OrdersCancelLineItemResponse Object}
\usage{
OrdersCancelLineItemResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersCancelLineItemResponse object
}
\description{
OrdersCancelLineItemResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
