% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequest}
\alias{OrdersCustomBatchRequest}
\title{OrdersCustomBatchRequest Object}
\usage{
OrdersCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
OrdersCustomBatchRequest object
}
\description{
OrdersCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersCustomBatchRequest functions: \code{\link{orders.custombatch}}
}

