% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntry}
\alias{OrdersCustomBatchRequestEntry}
\title{OrdersCustomBatchRequestEntry Object}
\usage{
OrdersCustomBatchRequestEntry(batchId = NULL, cancel = NULL,
  cancelLineItem = NULL, merchantId = NULL, merchantOrderId = NULL,
  method = NULL, operationId = NULL, orderId = NULL, refund = NULL,
  returnLineItem = NULL, shipLineItems = NULL, updateShipment = NULL)
}
\arguments{
\item{batchId}{An entry ID, unique within the batch request}

\item{cancel}{Required for cancel method}

\item{cancelLineItem}{Required for cancelLineItem method}

\item{merchantId}{The ID of the managing account}

\item{merchantOrderId}{The merchant order id}

\item{method}{The method to apply}

\item{operationId}{The ID of the operation}

\item{orderId}{The ID of the order}

\item{refund}{Required for refund method}

\item{returnLineItem}{Required for returnLineItem method}

\item{shipLineItems}{Required for shipLineItems method}

\item{updateShipment}{Required for updateShipment method}
}
\value{
OrdersCustomBatchRequestEntry object
}
\description{
OrdersCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
