% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryRefund}
\alias{OrdersCustomBatchRequestEntryRefund}
\title{OrdersCustomBatchRequestEntryRefund Object}
\usage{
OrdersCustomBatchRequestEntryRefund(amount = NULL, reason = NULL,
  reasonText = NULL)
}
\arguments{
\item{amount}{The amount that is refunded}

\item{reason}{The reason for the refund}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersCustomBatchRequestEntryRefund object
}
\description{
OrdersCustomBatchRequestEntryRefund Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
