% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersCustomBatchRequestEntryShipLineItems}
\alias{OrdersCustomBatchRequestEntryShipLineItems}
\title{OrdersCustomBatchRequestEntryShipLineItems Object}
\usage{
OrdersCustomBatchRequestEntryShipLineItems(carrier = NULL, lineItems = NULL,
  shipmentId = NULL, trackingId = NULL)
}
\arguments{
\item{carrier}{The carrier handling the shipment}

\item{lineItems}{Line items to ship}

\item{shipmentId}{The ID of the shipment}

\item{trackingId}{The tracking id for the shipment}
}
\value{
OrdersCustomBatchRequestEntryShipLineItems object
}
\description{
OrdersCustomBatchRequestEntryShipLineItems Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

