% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderShipment}
\alias{OrderShipment}
\title{OrderShipment Object}
\usage{
OrderShipment(carrier = NULL, creationDate = NULL, deliveryDate = NULL,
  id = NULL, lineItems = NULL, status = NULL, trackingId = NULL)
}
\arguments{
\item{carrier}{The carrier handling the shipment}

\item{creationDate}{Date on which the shipment has been created, in ISO 8601 format}

\item{deliveryDate}{Date on which the shipment has been delivered, in ISO 8601 format}

\item{id}{The id of the shipment}

\item{lineItems}{The line items that are shipped}

\item{status}{The status of the shipment}

\item{trackingId}{The tracking id for the shipment}
}
\value{
OrderShipment object
}
\description{
OrderShipment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

