% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrderShipmentLineItemShipment}
\alias{OrderShipmentLineItemShipment}
\title{OrderShipmentLineItemShipment Object}
\usage{
OrderShipmentLineItemShipment(lineItemId = NULL, quantity = NULL)
}
\arguments{
\item{lineItemId}{The id of the line item that is shipped}

\item{quantity}{The quantity that is shipped}
}
\value{
OrderShipmentLineItemShipment object
}
\description{
OrderShipmentLineItemShipment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

