% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersListResponse}
\alias{OrdersListResponse}
\title{OrdersListResponse Object}
\usage{
OrdersListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of orders}

\item{resources}{No description}
}
\value{
OrdersListResponse object
}
\description{
OrdersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
