% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersRefundRequest}
\alias{OrdersRefundRequest}
\title{OrdersRefundRequest Object}
\usage{
OrdersRefundRequest(amount = NULL, operationId = NULL, reason = NULL,
  reasonText = NULL)
}
\arguments{
\item{amount}{The amount that is refunded}

\item{operationId}{The ID of the operation}

\item{reason}{The reason for the refund}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersRefundRequest object
}
\description{
OrdersRefundRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersRefundRequest functions: \code{\link{orders.refund}}
}

