% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersRefundResponse}
\alias{OrdersRefundResponse}
\title{OrdersRefundResponse Object}
\usage{
OrdersRefundResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersRefundResponse object
}
\description{
OrdersRefundResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

