% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersReturnLineItemRequest}
\alias{OrdersReturnLineItemRequest}
\title{OrdersReturnLineItemRequest Object}
\usage{
OrdersReturnLineItemRequest(lineItemId = NULL, operationId = NULL,
  quantity = NULL, reason = NULL, reasonText = NULL)
}
\arguments{
\item{lineItemId}{The ID of the line item to return}

\item{operationId}{The ID of the operation}

\item{quantity}{The quantity to return}

\item{reason}{The reason for the return}

\item{reasonText}{The explanation of the reason}
}
\value{
OrdersReturnLineItemRequest object
}
\description{
OrdersReturnLineItemRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersReturnLineItemRequest functions: \code{\link{orders.returnlineitem}}
}
