% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersReturnLineItemResponse}
\alias{OrdersReturnLineItemResponse}
\title{OrdersReturnLineItemResponse Object}
\usage{
OrdersReturnLineItemResponse(executionStatus = NULL)
}
\arguments{
\item{executionStatus}{The status of the execution}
}
\value{
OrdersReturnLineItemResponse object
}
\description{
OrdersReturnLineItemResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

