% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{OrdersUpdateMerchantOrderIdRequest}
\alias{OrdersUpdateMerchantOrderIdRequest}
\title{OrdersUpdateMerchantOrderIdRequest Object}
\usage{
OrdersUpdateMerchantOrderIdRequest(merchantOrderId = NULL,
  operationId = NULL)
}
\arguments{
\item{merchantOrderId}{The merchant order id to be assigned to the order}

\item{operationId}{The ID of the operation}
}
\value{
OrdersUpdateMerchantOrderIdRequest object
}
\description{
OrdersUpdateMerchantOrderIdRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other OrdersUpdateMerchantOrderIdRequest functions: \code{\link{orders.updatemerchantorderid}}
}

