% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{PostalCodeGroup}
\alias{PostalCodeGroup}
\title{PostalCodeGroup Object}
\usage{
PostalCodeGroup(country = NULL, name = NULL, postalCodeRanges = NULL)
}
\arguments{
\item{country}{The CLDR territory code of the country the postal code group applies to}

\item{name}{The name of the postal code group, referred to in headers}

\item{postalCodeRanges}{A range of postal codes}
}
\value{
PostalCodeGroup object
}
\description{
PostalCodeGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
