% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{PostalCodeRange}
\alias{PostalCodeRange}
\title{PostalCodeRange Object}
\usage{
PostalCodeRange(postalCodeRangeBegin = NULL, postalCodeRangeEnd = NULL)
}
\arguments{
\item{postalCodeRangeBegin}{A postal code or a pattern of the form prefix* denoting the inclusive lower bound of the range defining the area}

\item{postalCodeRangeEnd}{A postal code or a pattern of the form prefix* denoting the inclusive upper bound of the range defining the area}
}
\value{
PostalCodeRange object
}
\description{
PostalCodeRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

