% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Price}
\alias{Price}
\title{Price Object}
\usage{
Price(currency = NULL, value = NULL)
}
\arguments{
\item{currency}{The currency of the price}

\item{value}{The price represented as a number}
}
\value{
Price object
}
\description{
Price Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

