% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Product}
\alias{Product}
\title{Product Object}
\usage{
Product(additionalImageLinks = NULL, additionalProductTypes = NULL,
  adult = NULL, adwordsGrouping = NULL, adwordsLabels = NULL,
  adwordsRedirect = NULL, ageGroup = NULL, aspects = NULL,
  availability = NULL, availabilityDate = NULL, brand = NULL,
  channel = NULL, color = NULL, condition = NULL,
  contentLanguage = NULL, customAttributes = NULL, customGroups = NULL,
  customLabel0 = NULL, customLabel1 = NULL, customLabel2 = NULL,
  customLabel3 = NULL, customLabel4 = NULL, description = NULL,
  destinations = NULL, displayAdsId = NULL, displayAdsLink = NULL,
  displayAdsSimilarIds = NULL, displayAdsTitle = NULL,
  displayAdsValue = NULL, energyEfficiencyClass = NULL,
  expirationDate = NULL, gender = NULL, googleProductCategory = NULL,
  gtin = NULL, id = NULL, identifierExists = NULL, imageLink = NULL,
  installment = NULL, isBundle = NULL, itemGroupId = NULL, link = NULL,
  loyaltyPoints = NULL, material = NULL, mobileLink = NULL, mpn = NULL,
  multipack = NULL, offerId = NULL, onlineOnly = NULL, pattern = NULL,
  price = NULL, productType = NULL, promotionIds = NULL,
  salePrice = NULL, salePriceEffectiveDate = NULL,
  sellOnGoogleQuantity = NULL, shipping = NULL, shippingHeight = NULL,
  shippingLabel = NULL, shippingLength = NULL, shippingWeight = NULL,
  shippingWidth = NULL, sizeSystem = NULL, sizeType = NULL,
  sizes = NULL, targetCountry = NULL, taxes = NULL, title = NULL,
  unitPricingBaseMeasure = NULL, unitPricingMeasure = NULL,
  validatedDestinations = NULL, warnings = NULL)
}
\arguments{
\item{additionalImageLinks}{Additional URLs of images of the item}

\item{additionalProductTypes}{Additional categories of the item (formatted as in products feed specification)}

\item{adult}{Set to true if the item is targeted towards adults}

\item{adwordsGrouping}{Used to group items in an arbitrary way}

\item{adwordsLabels}{Similar to adwords_grouping, but only works on CPC}

\item{adwordsRedirect}{Allows advertisers to override the item URL when the product is shown within the context of Product Ads}

\item{ageGroup}{Target age group of the item}

\item{aspects}{Specifies the intended aspects for the product}

\item{availability}{Availability status of the item}

\item{availabilityDate}{The day a pre-ordered product becomes available for delivery, in ISO 8601 format}

\item{brand}{Brand of the item}

\item{channel}{The item's channel (online or local)}

\item{color}{Color of the item}

\item{condition}{Condition or state of the item}

\item{contentLanguage}{The two-letter ISO 639-1 language code for the item}

\item{customAttributes}{A list of custom (merchant-provided) attributes}

\item{customGroups}{A list of custom (merchant-provided) custom attribute groups}

\item{customLabel0}{Custom label 0 for custom grouping of items in a Shopping campaign}

\item{customLabel1}{Custom label 1 for custom grouping of items in a Shopping campaign}

\item{customLabel2}{Custom label 2 for custom grouping of items in a Shopping campaign}

\item{customLabel3}{Custom label 3 for custom grouping of items in a Shopping campaign}

\item{customLabel4}{Custom label 4 for custom grouping of items in a Shopping campaign}

\item{description}{Description of the item}

\item{destinations}{Specifies the intended destinations for the product}

\item{displayAdsId}{An identifier for an item for dynamic remarketing campaigns}

\item{displayAdsLink}{URL directly to your item's landing page for dynamic remarketing campaigns}

\item{displayAdsSimilarIds}{Advertiser-specified recommendations}

\item{displayAdsTitle}{Title of an item for dynamic remarketing campaigns}

\item{displayAdsValue}{Offer margin for dynamic remarketing campaigns}

\item{energyEfficiencyClass}{The energy efficiency class as defined in EU directive 2010/30/EU}

\item{expirationDate}{Date on which the item should expire, as specified upon insertion, in ISO 8601 format}

\item{gender}{Target gender of the item}

\item{googleProductCategory}{Google's category of the item (see Google product taxonomy)}

\item{gtin}{Global Trade Item Number (GTIN) of the item}

\item{id}{The REST id of the product}

\item{identifierExists}{False when the item does not have unique product identifiers appropriate to its category, such as GTIN, MPN, and brand}

\item{imageLink}{URL of an image of the item}

\item{installment}{Number and amount of installments to pay for an item}

\item{isBundle}{Whether the item is a merchant-defined bundle}

\item{itemGroupId}{Shared identifier for all variants of the same product}

\item{link}{URL directly linking to your item's page on your website}

\item{loyaltyPoints}{Loyalty points that users receive after purchasing the item}

\item{material}{The material of which the item is made}

\item{mobileLink}{Link to a mobile-optimized version of the landing page}

\item{mpn}{Manufacturer Part Number (MPN) of the item}

\item{multipack}{The number of identical products in a merchant-defined multipack}

\item{offerId}{An identifier of the item}

\item{onlineOnly}{Whether an item is available for purchase only online}

\item{pattern}{The item's pattern (e}

\item{price}{Price of the item}

\item{productType}{Your category of the item (formatted as in products feed specification)}

\item{promotionIds}{The unique ID of a promotion}

\item{salePrice}{Advertised sale price of the item}

\item{salePriceEffectiveDate}{Date range during which the item is on sale (see products feed specification)}

\item{sellOnGoogleQuantity}{The quantity of the product that is reserved for sell-on-google ads}

\item{shipping}{Shipping rules}

\item{shippingHeight}{Height of the item for shipping}

\item{shippingLabel}{The shipping label of the product, used to group product in account-level shipping rules}

\item{shippingLength}{Length of the item for shipping}

\item{shippingWeight}{Weight of the item for shipping}

\item{shippingWidth}{Width of the item for shipping}

\item{sizeSystem}{System in which the size is specified}

\item{sizeType}{The cut of the item}

\item{sizes}{Size of the item}

\item{targetCountry}{The CLDR territory code for the item}

\item{taxes}{Tax information}

\item{title}{Title of the item}

\item{unitPricingBaseMeasure}{The preference of the denominator of the unit price}

\item{unitPricingMeasure}{The measure and dimension of an item}

\item{validatedDestinations}{The read-only list of intended destinations which passed validation}

\item{warnings}{Read-only warnings}
}
\value{
Product object
}
\description{
Product Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Product data.
}
\seealso{
Other Product functions: \code{\link{products.insert}}
}

