% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductAspect}
\alias{ProductAspect}
\title{ProductAspect Object}
\usage{
ProductAspect(aspectName = NULL, destinationName = NULL, intention = NULL)
}
\arguments{
\item{aspectName}{The name of the aspect}

\item{destinationName}{The name of the destination}

\item{intention}{Whether the aspect is required, excluded or should be validated}
}
\value{
ProductAspect object
}
\description{
ProductAspect Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
