% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductCustomAttribute}
\alias{ProductCustomAttribute}
\title{ProductCustomAttribute Object}
\usage{
ProductCustomAttribute(name = NULL, type = NULL, unit = NULL,
  value = NULL)
}
\arguments{
\item{name}{The name of the attribute}

\item{type}{The type of the attribute}

\item{unit}{Free-form unit of the attribute}

\item{value}{The value of the attribute}
}
\value{
ProductCustomAttribute object
}
\description{
ProductCustomAttribute Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
