% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductCustomGroup}
\alias{ProductCustomGroup}
\title{ProductCustomGroup Object}
\usage{
ProductCustomGroup(attributes = NULL, name = NULL)
}
\arguments{
\item{attributes}{The sub-attributes}

\item{name}{The name of the group}
}
\value{
ProductCustomGroup object
}
\description{
ProductCustomGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

