% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductDestination}
\alias{ProductDestination}
\title{ProductDestination Object}
\usage{
ProductDestination(destinationName = NULL, intention = NULL)
}
\arguments{
\item{destinationName}{The name of the destination}

\item{intention}{Whether the destination is required, excluded or should be validated}
}
\value{
ProductDestination object
}
\description{
ProductDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

