% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{products.list}
\alias{products.list}
\title{Lists the products in your Merchant Center account. This method can only be called for non-multi-client accounts.}
\usage{
products.list(merchantId, includeInvalidInsertedItems = NULL,
  maxResults = NULL, pageToken = NULL)
}
\arguments{
\item{merchantId}{The ID of the managing account}

\item{includeInvalidInsertedItems}{Flag to include the invalid inserted items in the result of the list request}

\item{maxResults}{The maximum number of products to return in the response, used for paging}

\item{pageToken}{The token returned by the previous request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}
}
