% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductsCustomBatchRequest}
\alias{ProductsCustomBatchRequest}
\title{ProductsCustomBatchRequest Object}
\usage{
ProductsCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
ProductsCustomBatchRequest object
}
\description{
ProductsCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProductsCustomBatchRequest functions: \code{\link{products.custombatch}}
}

