% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductsCustomBatchResponse}
\alias{ProductsCustomBatchResponse}
\title{ProductsCustomBatchResponse Object}
\usage{
ProductsCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
ProductsCustomBatchResponse object
}
\description{
ProductsCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
