% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductsCustomBatchResponseEntry}
\alias{ProductsCustomBatchResponseEntry}
\title{ProductsCustomBatchResponseEntry Object}
\usage{
ProductsCustomBatchResponseEntry(batchId = NULL, errors = NULL,
  product = NULL)
}
\arguments{
\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors defined if and only if the request failed}

\item{product}{The inserted product}
}
\value{
ProductsCustomBatchResponseEntry object
}
\description{
ProductsCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch products response.
}
