% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductShipping}
\alias{ProductShipping}
\title{ProductShipping Object}
\usage{
ProductShipping(country = NULL, locationGroupName = NULL,
  locationId = NULL, postalCode = NULL, price = NULL, region = NULL,
  service = NULL)
}
\arguments{
\item{country}{The CLDR territory code of the country to which an item will ship}

\item{locationGroupName}{The location where the shipping is applicable, represented by a location group name}

\item{locationId}{The numeric id of a location that the shipping rate applies to as defined in the AdWords API}

\item{postalCode}{The postal code range that the shipping rate applies to, represented by a postal code, a postal code prefix followed by a * wildcard, a range between two postal codes or two postal code prefixes of equal length}

\item{price}{Fixed shipping price, represented as a number}

\item{region}{The geographic region to which a shipping rate applies (e}

\item{service}{A free-form description of the service class or delivery speed}
}
\value{
ProductShipping object
}
\description{
ProductShipping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

