% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductShippingDimension}
\alias{ProductShippingDimension}
\title{ProductShippingDimension Object}
\usage{
ProductShippingDimension(unit = NULL, value = NULL)
}
\arguments{
\item{unit}{The unit of value}

\item{value}{The dimension of the product used to calculate the shipping cost of the item}
}
\value{
ProductShippingDimension object
}
\description{
ProductShippingDimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
