% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductShippingWeight}
\alias{ProductShippingWeight}
\title{ProductShippingWeight Object}
\usage{
ProductShippingWeight(unit = NULL, value = NULL)
}
\arguments{
\item{unit}{The unit of value}

\item{value}{The weight of the product used to calculate the shipping cost of the item}
}
\value{
ProductShippingWeight object
}
\description{
ProductShippingWeight Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
