% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductsListResponse}
\alias{ProductsListResponse}
\title{ProductsListResponse Object}
\usage{
ProductsListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of products}

\item{resources}{No description}
}
\value{
ProductsListResponse object
}
\description{
ProductsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

