% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductStatus}
\alias{ProductStatus}
\title{ProductStatus Object}
\usage{
ProductStatus(creationDate = NULL, dataQualityIssues = NULL,
  destinationStatuses = NULL, googleExpirationDate = NULL,
  lastUpdateDate = NULL, link = NULL, productId = NULL, title = NULL)
}
\arguments{
\item{creationDate}{Date on which the item has been created, in ISO 8601 format}

\item{dataQualityIssues}{A list of data quality issues associated with the product}

\item{destinationStatuses}{The intended destinations for the product}

\item{googleExpirationDate}{Date on which the item expires in Google Shopping, in ISO 8601 format}

\item{lastUpdateDate}{Date on which the item has been last updated, in ISO 8601 format}

\item{link}{The link to the product}

\item{productId}{The id of the product for which status is reported}

\item{title}{The title of the product}
}
\value{
ProductStatus object
}
\description{
ProductStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The status of a product, i.e., information about a product computed asynchronously by the data quality analysis.
}

