% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductStatusDataQualityIssue}
\alias{ProductStatusDataQualityIssue}
\title{ProductStatusDataQualityIssue Object}
\usage{
ProductStatusDataQualityIssue(detail = NULL, fetchStatus = NULL,
  id = NULL, location = NULL, severity = NULL, timestamp = NULL,
  valueOnLandingPage = NULL, valueProvided = NULL)
}
\arguments{
\item{detail}{A more detailed error string}

\item{fetchStatus}{The fetch status for landing_page_errors}

\item{id}{The id of the data quality issue}

\item{location}{The attribute name that is relevant for the issue}

\item{severity}{The severity of the data quality issue}

\item{timestamp}{The time stamp of the data quality issue}

\item{valueOnLandingPage}{The value of that attribute that was found on the landing page}

\item{valueProvided}{The value the attribute had at time of evaluation}
}
\value{
ProductStatusDataQualityIssue object
}
\description{
ProductStatusDataQualityIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

