% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductStatusDestinationStatus}
\alias{ProductStatusDestinationStatus}
\title{ProductStatusDestinationStatus Object}
\usage{
ProductStatusDestinationStatus(approvalStatus = NULL, destination = NULL,
  intention = NULL)
}
\arguments{
\item{approvalStatus}{The destination's approval status}

\item{destination}{The name of the destination}

\item{intention}{Whether the destination is required, excluded, selected by default or should be validated}
}
\value{
ProductStatusDestinationStatus object
}
\description{
ProductStatusDestinationStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

