% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductstatusesCustomBatchRequest}
\alias{ProductstatusesCustomBatchRequest}
\title{ProductstatusesCustomBatchRequest Object}
\usage{
ProductstatusesCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
ProductstatusesCustomBatchRequest object
}
\description{
ProductstatusesCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ProductstatusesCustomBatchRequest functions: \code{\link{productstatuses.custombatch}}
}
