% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductstatusesCustomBatchRequestEntry}
\alias{ProductstatusesCustomBatchRequestEntry}
\title{ProductstatusesCustomBatchRequestEntry Object}
\usage{
ProductstatusesCustomBatchRequestEntry(batchId = NULL, merchantId = NULL,
  method = NULL, productId = NULL)
}
\arguments{
\item{batchId}{An entry ID, unique within the batch request}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}

\item{productId}{The ID of the product whose status to get}
}
\value{
ProductstatusesCustomBatchRequestEntry object
}
\description{
ProductstatusesCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch productstatuses request.
}
