% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductstatusesCustomBatchResponse}
\alias{ProductstatusesCustomBatchResponse}
\title{ProductstatusesCustomBatchResponse Object}
\usage{
ProductstatusesCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
ProductstatusesCustomBatchResponse object
}
\description{
ProductstatusesCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
