% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductstatusesCustomBatchResponseEntry}
\alias{ProductstatusesCustomBatchResponseEntry}
\title{ProductstatusesCustomBatchResponseEntry Object}
\usage{
ProductstatusesCustomBatchResponseEntry(batchId = NULL, errors = NULL,
  productStatus = NULL)
}
\arguments{
\item{batchId}{The ID of the request entry this entry responds to}

\item{errors}{A list of errors, if the request failed}

\item{productStatus}{The requested product status}
}
\value{
ProductstatusesCustomBatchResponseEntry object
}
\description{
ProductstatusesCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch productstatuses response.
}

