% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductstatusesListResponse}
\alias{ProductstatusesListResponse}
\title{ProductstatusesListResponse Object}
\usage{
ProductstatusesListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of products statuses}

\item{resources}{No description}
}
\value{
ProductstatusesListResponse object
}
\description{
ProductstatusesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

