% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductTax}
\alias{ProductTax}
\title{ProductTax Object}
\usage{
ProductTax(country = NULL, locationId = NULL, postalCode = NULL,
  rate = NULL, region = NULL, taxShip = NULL)
}
\arguments{
\item{country}{The country within which the item is taxed, specified as a CLDR territory code}

\item{locationId}{The numeric id of a location that the tax rate applies to as defined in the AdWords API}

\item{postalCode}{The postal code range that the tax rate applies to, represented by a ZIP code, a ZIP code prefix using * wildcard, a range between two ZIP codes or two ZIP code prefixes of equal length}

\item{rate}{The percentage of tax rate that applies to the item price}

\item{region}{The geographic region to which the tax rate applies}

\item{taxShip}{Set to true if tax is charged on shipping}
}
\value{
ProductTax object
}
\description{
ProductTax Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

