% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductUnitPricingBaseMeasure}
\alias{ProductUnitPricingBaseMeasure}
\title{ProductUnitPricingBaseMeasure Object}
\usage{
ProductUnitPricingBaseMeasure(unit = NULL, value = NULL)
}
\arguments{
\item{unit}{The unit of the denominator}

\item{value}{The denominator of the unit price}
}
\value{
ProductUnitPricingBaseMeasure object
}
\description{
ProductUnitPricingBaseMeasure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
