% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ProductUnitPricingMeasure}
\alias{ProductUnitPricingMeasure}
\title{ProductUnitPricingMeasure Object}
\usage{
ProductUnitPricingMeasure(unit = NULL, value = NULL)
}
\arguments{
\item{unit}{The unit of the measure}

\item{value}{The measure of an item}
}
\value{
ProductUnitPricingMeasure object
}
\description{
ProductUnitPricingMeasure Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

