% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{RateGroup}
\alias{RateGroup}
\title{RateGroup Object}
\usage{
RateGroup(applicableShippingLabels = NULL, carrierRates = NULL,
  mainTable = NULL, singleValue = NULL, subtables = NULL)
}
\arguments{
\item{applicableShippingLabels}{A list of shipping labels defining the products to which this rate group applies to}

\item{carrierRates}{A list of carrier rates that can be referred to by mainTable or singleValue}

\item{mainTable}{A table defining the rate group, when singleValue is not expressive enough}

\item{singleValue}{The value of the rate group (e}

\item{subtables}{A list of subtables referred to by mainTable}
}
\value{
RateGroup object
}
\description{
RateGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

