% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Row}
\alias{Row}
\title{Row Object}
\usage{
Row(cells = NULL)
}
\arguments{
\item{cells}{The list of cells that constitute the row}
}
\value{
Row object
}
\description{
Row Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
