% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Service}
\alias{Service}
\title{Service Object}
\usage{
Service(active = NULL, currency = NULL, deliveryCountry = NULL,
  deliveryTime = NULL, name = NULL, rateGroups = NULL)
}
\arguments{
\item{active}{A boolean exposing the active status of the shipping service}

\item{currency}{The CLDR code of the currency to which this service applies}

\item{deliveryCountry}{The CLDR territory code of the country to which the service applies}

\item{deliveryTime}{Time spent in various aspects from order to the delivery of the product}

\item{name}{Free-form name of the service}

\item{rateGroups}{Shipping rate group definitions}
}
\value{
Service object
}
\description{
Service Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

