% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_functions.R
\name{shippingsettings.custombatch}
\alias{shippingsettings.custombatch}
\title{Retrieves and updates the shipping settings of multiple accounts in a single request.}
\usage{
shippingsettings.custombatch(ShippingsettingsCustomBatchRequest,
  dryRun = NULL)
}
\arguments{
\item{ShippingsettingsCustomBatchRequest}{The \link{ShippingsettingsCustomBatchRequest} object to pass to this method}

\item{dryRun}{Flag to run the request in dry-run mode}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/content
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/content)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/shopping-content}{Google Documentation}

Other ShippingsettingsCustomBatchRequest functions: \code{\link{ShippingsettingsCustomBatchRequest}}
}
