% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingSettings}
\alias{ShippingSettings}
\title{ShippingSettings Object}
\usage{
ShippingSettings(accountId = NULL, postalCodeGroups = NULL,
  services = NULL)
}
\arguments{
\item{accountId}{The ID of the account to which these account shipping settings belong}

\item{postalCodeGroups}{A list of postal code groups that can be referred to in services}

\item{services}{The target account's list of services}
}
\value{
ShippingSettings object
}
\description{
ShippingSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The merchant account's shipping settings.
}
\seealso{
Other ShippingSettings functions: \code{\link{shippingsettings.patch}},
  \code{\link{shippingsettings.update}}
}

