% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsCustomBatchRequest}
\alias{ShippingsettingsCustomBatchRequest}
\title{ShippingsettingsCustomBatchRequest Object}
\usage{
ShippingsettingsCustomBatchRequest(entries = NULL)
}
\arguments{
\item{entries}{The request entries to be processed in the batch}
}
\value{
ShippingsettingsCustomBatchRequest object
}
\description{
ShippingsettingsCustomBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ShippingsettingsCustomBatchRequest functions: \code{\link{shippingsettings.custombatch}}
}
