% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsCustomBatchRequestEntry}
\alias{ShippingsettingsCustomBatchRequestEntry}
\title{ShippingsettingsCustomBatchRequestEntry Object}
\usage{
ShippingsettingsCustomBatchRequestEntry(accountId = NULL, batchId = NULL,
  merchantId = NULL, method = NULL, shippingSettings = NULL)
}
\arguments{
\item{accountId}{The ID of the account for which to get/update account shipping settings}

\item{batchId}{An entry ID, unique within the batch request}

\item{merchantId}{The ID of the managing account}

\item{method}{No description}

\item{shippingSettings}{The account shipping settings to update}
}
\value{
ShippingsettingsCustomBatchRequestEntry object
}
\description{
ShippingsettingsCustomBatchRequestEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch accountshipping request.
}

