% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsCustomBatchResponse}
\alias{ShippingsettingsCustomBatchResponse}
\title{ShippingsettingsCustomBatchResponse Object}
\usage{
ShippingsettingsCustomBatchResponse(entries = NULL)
}
\arguments{
\item{entries}{The result of the execution of the batch requests}
}
\value{
ShippingsettingsCustomBatchResponse object
}
\description{
ShippingsettingsCustomBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
