% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsCustomBatchResponseEntry}
\alias{ShippingsettingsCustomBatchResponseEntry}
\title{ShippingsettingsCustomBatchResponseEntry Object}
\usage{
ShippingsettingsCustomBatchResponseEntry(batchId = NULL, errors = NULL,
  shippingSettings = NULL)
}
\arguments{
\item{batchId}{The ID of the request entry to which this entry responds}

\item{errors}{A list of errors defined if, and only if, the request failed}

\item{shippingSettings}{The retrieved or updated account shipping settings}
}
\value{
ShippingsettingsCustomBatchResponseEntry object
}
\description{
ShippingsettingsCustomBatchResponseEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A batch entry encoding a single non-batch shipping settings response.
}

