% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsGetSupportedCarriersResponse}
\alias{ShippingsettingsGetSupportedCarriersResponse}
\title{ShippingsettingsGetSupportedCarriersResponse Object}
\usage{
ShippingsettingsGetSupportedCarriersResponse(carriers = NULL)
}
\arguments{
\item{carriers}{A list of supported carriers}
}
\value{
ShippingsettingsGetSupportedCarriersResponse object
}
\description{
ShippingsettingsGetSupportedCarriersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
