% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{ShippingsettingsListResponse}
\alias{ShippingsettingsListResponse}
\title{ShippingsettingsListResponse Object}
\usage{
ShippingsettingsListResponse(nextPageToken = NULL, resources = NULL)
}
\arguments{
\item{nextPageToken}{The token for the retrieval of the next page of shipping settings}

\item{resources}{No description}
}
\value{
ShippingsettingsListResponse object
}
\description{
ShippingsettingsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

