% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Table}
\alias{Table}
\title{Table Object}
\usage{
Table(columnHeaders = NULL, name = NULL, rowHeaders = NULL, rows = NULL)
}
\arguments{
\item{columnHeaders}{Headers of the table's columns}

\item{name}{Name of the table}

\item{rowHeaders}{Headers of the table's rows}

\item{rows}{The list of rows that constitute the table}
}
\value{
Table object
}
\description{
Table Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
