% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{TestOrder}
\alias{TestOrder}
\title{TestOrder Object}
\usage{
TestOrder(customer = NULL, lineItems = NULL, paymentMethod = NULL,
  predefinedDeliveryAddress = NULL, promotions = NULL,
  shippingCost = NULL, shippingCostTax = NULL, shippingOption = NULL)
}
\arguments{
\item{customer}{The details of the customer who placed the order}

\item{lineItems}{Line items that are ordered}

\item{paymentMethod}{The details of the payment method}

\item{predefinedDeliveryAddress}{Identifier of one of the predefined delivery addresses for the delivery}

\item{promotions}{The details of the merchant provided promotions applied to the order}

\item{shippingCost}{The total cost of shipping for all items}

\item{shippingCostTax}{The tax for the total shipping cost}

\item{shippingOption}{The requested shipping option}
}
\value{
TestOrder object
}
\description{
TestOrder Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
