% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_objects.R
\name{Weight}
\alias{Weight}
\title{Weight Object}
\usage{
Weight(unit = NULL, value = NULL)
}
\arguments{
\item{unit}{The weight unit}

\item{value}{The weight represented as a number}
}
\value{
Weight object
}
\description{
Weight Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
